fix a hashcode & equals contract
Due to the fact that the previous implementation might lead to the hashcode collision while used within Map collections it's required to follow the proper equals and hashcode contract. All the fields of underlying classes must be included in the hashcode and equals generation.

Having the previous implementation it's fairly possible that several instances of the "Client" or any other entity would be included in the Map as a key which would lead to the hashcode collision and those would be put into the same bucket, therefore, leading to lower performance.

See: https://www.baeldung.com/java-equals-hashcode-contracts